Use microsecond precision for last modification time
When the file system allows it, this prevents Tumbler from believing
that the original file has not been modified since the last time the
thumbnail was created, when in fact modifications have taken place in
the second that the thumbnail was created.

The time of last modification is stored internally in a `gdouble`, and
written to the PNG thumbnail in `%.6f` format. This complies well with
the Freedesktop specification [1], as the stat command returns this
format when invoked as `stat -c%.6Y file`, although it is likely that
the specification was written with the idea that this value be an
integer.

The extraction of the thumbnail information is done via
`g_ascii_strtod()`, so an integer or less precision is not a problem. As
for a code expecting to find an integer, it is likely to simply ignore
the decimal part, as Tumbler used to do by extracting the information
via `atol()`. It is possible, however, that an error will be found if a
more complete extractor is used, such as `strtol()`.

The changes this causes to the exposed Tumbler APIs should also be
inconsequential, as it is a conversion to a higher type. This should
therefore result in implicit conversions, except perhaps for
`tumbler_file_info_get_mtime()`, if its return value is explicitly
stored in a `gint64`.

[1] https://specifications.freedesktop.org/thumbnail-spec/thumbnail-spec-latest.html#CREATION

Fixes #15, see !20 for more details.

+++ plugins/xdg-cache/xdg-cache-cache.h
@@ -46,12 +46,12 @@ GFile   *xdg_cache_cache_get_temp_file        (const gchar            *uri,
                                                TumblerThumbnailFlavor  *flavor) G_GNUC_MALLOC G_GNUC_WARN_UNUSED_RESULT;
 gboolean xdg_cache_cache_read_thumbnail_info  (const gchar            *filename,
                                                gchar                 **uri,
-                                               guint64                *mtime,
+                                               gdouble                *mtime,
                                                GCancellable           *cancellable,
                                                GError                **error);
 gboolean xdg_cache_cache_write_thumbnail_info (const gchar            *filename,
                                                const gchar            *uri,
-                                               guint64                 mtime,
+                                               gdouble                 mtime,
                                                GCancellable           *cancellable,
                                                GError                **error);
